requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2.7
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
oauthlib==3.2.2
packaging==24.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
whitenoise==6.6.0
